refactor: split concept strip view
