+++ public/index.html
@@ -462,7 +462,7 @@
   <div id="tooltip-root" aria-hidden="true"></div>
 
   <script src="js/drill-chamber.js?v=6"></script>
-  <script type="module" src="js/app.js?v=106"></script>
+  <script type="module" src="js/app.js?v=107"></script>
   <script src="js/intro-particles.js?v=8"></script>
   <script type="module" src="js/tooltips.js?v=4"></script>
   <script type="module" src="js/floating-room-label.js?v=5"></script>
